+++ agave_app/python/pyrenderer.h
@@ -9,7 +9,7 @@
 #include "renderlib/gfxOpenGL/RendererGLContext.h"
 #include "renderlib/graphics/IRenderWindow.h"
 #include "renderlib/graphics/GestureGraphicsGL.h"
-#include "renderlib/graphics/gl/Util.h"
+#include "renderlib/gfxOpenGL/Util.h"
 #include "renderlib/renderlib.h"
 
 #include <QList>
