+++ renderlib/gfxOpenGL/CMakeLists.txt
@@ -9,12 +9,22 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Backend.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Device.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Device.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLContext.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLContext.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/HeadlessGLContext.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/HeadlessGLContext.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RendererGLContext.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RendererGLContext.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Util.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Util.h"
 )
 
 target_link_libraries(renderlib
