+++ renderlib/gfxOpenGL/Image3D.cpp
@@ -1,11 +1,11 @@
 
-#include "gl/Image3D.h"
+#include "Image3D.h"
 
 #include "Fuse.h"
 #include "ImageXYZC.h"
 #include "Logging.h"
 #include "RenderSettings.h"
-#include "gl/Util.h"
+#include "Util.h"
 
 #include <algorithm>
 #include <array>
