+++ renderlib/graphics/RenderGLPT.cpp
@@ -6,9 +6,9 @@
 #include "Framebuffer.h"
 #include "ImageXYZC.h"
 #include "Logging.h"
-#include "gl/FSQ.h"
-#include "gl/Image3D.h"
-#include "gl/Util.h"
+#include "gfxOpenGL/FSQ.h"
+#include "gfxOpenGL/Image3D.h"
+#include "gfxOpenGL/Util.h"
 #include "glsl/GLCopyShader.h"
 #include "glsl/GLImageShader2DnoLut.h"
 #include "glsl/GLPTVolumeShader.h"
