--- renderlib/graphics/gl/CMakeLists.txt
@@ -1,15 +1,0 @@
-target_include_directories(renderlib PUBLIC
-	"${CMAKE_CURRENT_SOURCE_DIR}"
-)
-target_sources(renderlib PRIVATE
-	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FontGL.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Image2D.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Util.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Util.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FSQ.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/Image3D.h"
-)
