+++ renderlib/graphics/glsl/GLImageShader2DnoLut.cpp
@@ -3,7 +3,7 @@
 #include "Logging.h"
 #include "shaders.h"
 
-#include <gl/Util.h>
+#include "gfxOpenGL/Util.h"
 #include <glm.h>
 
 #include <iostream>
