+++ renderlib/graphics/glsl/GLImageShader2DnoLut.h
@@ -2,7 +2,7 @@
 
 #include "gfxOpenGL/Backend.h"
 
-#include "gl/Util.h"
+#include "gfxOpenGL/Util.h"
 
 #include <glm.h>
