+++ renderlib/graphics/glsl/GLPTVolumeShader.cpp
@@ -9,7 +9,7 @@
 #include "Logging.h"
 #include "shaders.h"
 
-#include <gl/Util.h>
+#include "gfxOpenGL/Util.h"
 #include <glm.h>
 
 #include <iostream>
