+++ renderlib/graphics/glsl/GLToneMapShader.cpp
@@ -5,7 +5,7 @@
 #include "Logging.h"
 #include "shaders.h"
 
-#include <gl/Util.h>
+#include "gfxOpenGL/Util.h"
 #include <glm.h>
 
 #include <iostream>
